clean up getMathUnit and rename file.
